refactor: setImage 함수 이동
