+++ app/src/main/java/woowacourse/movie/ui/view/utils/ImageViewExtension.kt
@@ -1,0 +1,13 @@
+package woowacourse.movie.ui.view.utils
+
+import android.view.View
+import android.widget.ImageView
+
+fun ImageView.setImage(res: Int?) {
+    if (res != null) {
+        this.setImageResource(res)
+        this.visibility = View.VISIBLE
+    } else {
+        this.visibility = View.INVISIBLE
+    }
+}
